Кэширование данных полученных методами analytics с помощью базы данных redis
